src/activity-service/pyproject.toml
[build-system]
requires = ["poetry>=2.0.1"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
name = "serverless_sample_app"
version = "3.0.0"
authors = ["DataDog"]
classifiers=[
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.13",
]
repository="https://github.com/DataDog/serverless-sample-app"
readme = "README.md"
license = "MIT-0"

[tool.poetry.dependencies]
python = "^3.13.0"
pydantic = {version = "^2.0.3"}
email-validator = {version = "*"}
aws-lambda-powertools =  {extras = ["tracer"],version = "^3.15.1"}
mypy-boto3-dynamodb = "*"
cachetools = "*"
boto3 = "^1.38.41"
aws-lambda-env-modeler = "*"

[tool.poetry.group.dev.dependencies]
# CDK
activity-service-cdk = {path = "cdk", develop = true}
aws-cdk-lib = ">=2.201.0"
datadog-cdk-constructs-v2 = "^2.6.0"
constructs = ">=10.0.0"
cdk-nag = ">2.0.0"
"aws-cdk.aws-lambda-python-alpha" = "^2.201.0-alpha.0"
cdk-monitoring-constructs = "*"
# DEV
pytest = "*"
pytest-mock = "*"
pycodestyle = "*"
pytest-cov = "*"
pytest-html = "*"
python-dateutil = "*"
python-dotenv = "*"
GitPython = "*"
ruff = "*"
radon = "*"
xenon = "*"
pre-commit = "*"
mkdocs-material = "*"
mkdocs-git-revision-date-plugin = "*"
mkdocs-glightbox = "*"
types-cachetools = "*"
mypy = "*"
types-requests = "*"
toml = "*"
mkdocs-render-swagger-plugin = "*"
boto3 = "^1.26.125"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.9"

[tool.pytest.ini_options]
testpaths = "tests"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "cdk.out",
    ".vscode",
    ".pytest_cache",
    ".build",
    ".mypy_cache",
    ".github"
]

# Same as Black.
line-length = 150
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [ "E203", "E266", "E501", "W191"]

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
known-third-party = ["pydantic", "aws_lambda_powertools"]
